fixed game not working under Firefox

+++ public/shader/base.vert
@@ -7,7 +7,7 @@ uniform bool uHasColor;
 uniform bool uHasNormal;
 
 uniform bool uHasFog;
-uniform int uFogMode;
+uniform highp int uFogMode;
 uniform highp vec4 uFogColor;
 uniform highp vec2 uFogPosition;
 uniform highp float uFogDensity;
